Got translations working.

+++ build/make.d/etc.bash
@@ -3,6 +3,7 @@ echo -e "\e[34;1m::\e[0;1m Deploying server configuration...\e[0m"
 cd ../../etc
 ## Have to use $(ls -A) in order to copy hidden files.
 for F in $($(which ls) -A); do
+	[[ "$F" == '.keep' ]] && continue
 	cp -afrv "$F" /etc/ | sed 's/^.*[-][>] .//gm' | sed 's/.$//gm'
 done
 echo -e "\e[34;1m::\e[0;1m Done.\e[0m"
